% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iroc.R
\name{iroc}
\alias{iroc}
\title{Plot the ROC curve}
\usage{
iroc(testy, yhat)
}
\arguments{
\item{testy}{A numeric vector containing  prediction probabilities.}

\item{yhat}{A vector containing the true class labels.}
}
\value{
The figure of the ROC curve.
}
\description{
Plot the ROC curve and show the AUC of ROC.
The 0.95 CI of AUC were calculated by DeLong's method.
}
\examples{
#Simulate true prediction labels
testy<-rep(c(0,1),500)
#Simulate prediction probabilities
yhat<-runif(1000, min = 0, max = 1)
#Plot the ROC curve and show the AUC and its CI
iroc(testy, yhat)
}
\author{
Yi Li, \email{liyistat@gmail.com}
}

