% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iroc.R
\name{perf}
\alias{perf}
\title{Calculate the performance for the ROC curve}
\usage{
perf(testy, yhat)
}
\arguments{
\item{testy}{A vector containing the true class labels.}

\item{yhat}{A numeric vector containing  prediction probabilities.}
}
\value{
Performance for the ROC curve
}
\description{
Calculate the performance for the ROC curve
}

