% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iroc.R
\name{iroc}
\alias{iroc}
\title{Plot the ROC curve}
\usage{
iroc(testy, yhat)
}
\arguments{
\item{testy}{A numeric vector containing  prediction probabilities.}

\item{yhat}{A vector containing the true class labels.}
}
\value{
The figure of the ROC curve.
}
\description{
Plot the ROC curve and show the AUC of ROC.
The 0.95 CI of AUC were calculated by DeLong's method.
}

