% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iroc.R
\name{pred.result}
\alias{pred.result}
\title{Calculate sensitivity, specitifity, accuracy and AUC of the ROC.}
\usage{
pred.result(testy, yhat)
}
\arguments{
\item{testy}{A numeric vector containing  prediction probabilities.}

\item{yhat}{A vector containing the true class labels.}
}
\value{
result A list of sensitivity, specifity, accuracy, AUC and 0.95 CI of AUC.
}
\description{
Calculate sensitivity, specitifity, accuracy and AUC of the ROC.
And the 0.95 CI of AUC were calculated by DeLong's method.
}

